Reorganize builders
- Moves contents of `statistics` to `builders`
- Moves base_table_builder from root to builders
