+++ cumulus_library/__init__.py
@@ -1,8 +1,8 @@
 """Package metadata"""
 
-from .base_table_builder import BaseTableBuilder
 from .base_utils import StudyConfig
-from .statistics.counts import CountsBuilder
+from .builders.base_table_builder import BaseTableBuilder
+from .builders.counts import CountsBuilder
 from .study_manifest import StudyManifest
 
 __all__ = ["BaseTableBuilder", "CountsBuilder", "StudyConfig", "StudyManifest"]
