+++ cumulus_library/builders/counts.py
@@ -3,14 +3,14 @@
 import sys
 from pathlib import Path
 
-from cumulus_library import base_table_builder, errors, study_manifest
+from cumulus_library import BaseTableBuilder, errors, study_manifest
 from cumulus_library.statistics.statistics_templates import counts_templates
 
 # Defined here for easy overriding by tests
 DEFAULT_MIN_SUBJECT = 10
 
 
-class CountsBuilder(base_table_builder.BaseTableBuilder):
+class CountsBuilder(BaseTableBuilder):
     """Extends BaseTableBuilder for counts-related use cases"""
 
     def __init__(self, study_prefix: str | None = None):
