+++ cumulus_library/builders/statistics_templates/count.sql.jinja
