+++ cumulus_library/builders/statistics_templates/psm_create_covariate_table.sql.jinja
