+++ cumulus_library/statistics/README.md
@@ -1,0 +1,5 @@
+# Deprecation notice
+
+The statistics module has been combined with other tables builders into
+the new `builders` module. Backwards compatibility for importing from 
+`statistics` will be removed in a future version.
