+++ cumulus_library/statistics/statistics_templates/__init__.py
@@ -1,0 +1,1 @@
+from cumulus_library.builders.statistics_templates import counts_templates, psm_templates
